角色权限管理后台及部分页面样式改版

+++ src/main/java/warmer/star/blog/util/AppUserUtil.java
@@ -41,8 +41,8 @@ public class AppUserUtil {
 		return result;
 	}
 
-	public static long GetUserId() {
-		long result = 0;
+	public static Integer GetUserId() {
+		Integer result = 0;
 
 		try {
 			if (IsUserExsit()) {
